[tasks] Factor out `TaskDetailContainer`
Prepare for more responsive design to utilize more window space
