+++ packages/tasks/src/components/includes/TaskDetailPanel.module.css
@@ -24,15 +24,3 @@
     width: 500px !important;
   }
 }
-
-.TaskDetailTitle {
-  padding: 1em;
-}
-
-.TaskDetailActions {
-  padding: 1em;
-}
-
-.TaskDetailActionElement {
-  margin: 0 0.5em !important;
-}
